Safari-609.1.11.1/LayoutTests/streams/reference-implementation/pipe-to-options-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: destination error
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: destination error
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: destination error
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: destination error

Harness Error (FAIL), message = destination error

PASS Piping with no options and a destination error 
PASS Piping with { preventCancel: false } and a destination error 
PASS Piping with { preventCancel: true } and a destination error 
